suchigoapi/api/requirements.txt
Django>=5.1.5
djangorestframework>=3.14.0
Pillow>=10.0.0